feat: Add support for inserting variables in group merge messages
This commit adds support for inserting variables in group merge messages. It introduces a new `PublicNameTag` component that allows users to insert the `{{public_name}}` variable into the message text. The `insertVariablesToMessage` function is used to insert the variable at the current cursor position in the message textarea. This feature enhances the customization and personalization of group merge messages.