refactor: Add multiple_responses property to MergedGroup
The code changes in this commit add the `multiple_responses` property to the `MergedGroup` type in both the client and server code. This property allows for tracking whether a merged group can have multiple responses or not. By default, the `multiple_responses` property is set to `false`. This change provides flexibility in handling merged groups with different response requirements.