refactor: Add forward functionality to MergeGroupValidator
The MergeGroupValidator function in groups.validator.ts has been updated to include a new forward property in the MergeGroupValidationResult type. This property allows for forwarding messages to another number when certain conditions are met. This refactor enhances the functionality of the MergeGroupValidator by providing the ability to forward messages to another number.